AH: refactor useGames hook using react query and delete(commented) useData file
