+++ src/components/GameCard.tsx
@@ -15,10 +15,9 @@ import Emoji from "./Emoji";
 
 interface Props {
   game: Game;
-  index: number;
 }
 
-const GameCard = ({ game, index }: Props) => {
+const GameCard = ({ game }: Props) => {
   const cardBackground = useColorModeValue("gray.100", "blackAlpha.600");
   return (
     <Card bg={cardBackground}>
