+++ src/components/GameGrid.tsx
@@ -26,7 +26,7 @@ const GameGrid = ({ gameQuery }: Props) => {
         // TODO: remove hard coded width
         <Alert status="error" width="65%">
           <AlertIcon></AlertIcon>
-          <AlertTitle>{error}</AlertTitle>
+          <AlertTitle>{error.message}</AlertTitle>
           <AlertDescription>{ERROR_FETCHING_DATA}</AlertDescription>
         </Alert>
       ) : (
@@ -42,7 +42,7 @@ const GameGrid = ({ gameQuery }: Props) => {
                 <GameCardSkeleton></GameCardSkeleton>
               </GameCardContainer>
             ))}
-          {data.map((game, index) => (
+          {data?.results.map((game, index) => (
             <GameCardContainer key={game.id}>
               <GameCard game={game} index={index}></GameCard>
             </GameCardContainer>
