+++ src/hooks/useData.ts
@@ -1,49 +1,42 @@
-import { useEffect, useState } from "react";
-import apiClient from "../services/api-client";
-import { AxiosRequestConfig, CanceledError } from "axios";
+// No longer need useData file - replaced with react query to fetch data from backend
 
-export interface FetchResponse<T> {
-  count: number;
-  results: T[];
-}
+// const useData = <T>(
+//   endpoint: string,
+//   requestConfig?: AxiosRequestConfig,
+//   deps?: any[]
+// ) => {
+//   const [data, setData] = useState<T[]>([]);
+//   const [error, setError] = useState("");
+//   const [isLoading, setLoading] = useState(false);
 
-const useData = <T>(
-  endpoint: string,
-  requestConfig?: AxiosRequestConfig,
-  deps?: any[]
-) => {
-  const [data, setData] = useState<T[]>([]);
-  const [error, setError] = useState("");
-  const [isLoading, setLoading] = useState(false);
+//   useEffect(
+//     () => {
+//       const controller = new AbortController();
 
-  useEffect(
-    () => {
-      const controller = new AbortController();
+//       setLoading(true);
+//       apiClient
+//         .get<FetchResponse<T>>(endpoint, {
+//           signal: controller.signal,
+//           ...requestConfig,
+//         })
+//         .then((response) => {
+//           setData(response.data.results);
+//           setLoading(false);
+//           console.log(response.data.results);
+//         })
+//         .catch((error) => {
+//           if (error instanceof CanceledError) {
+//             return;
+//           }
+//           setError(error.message + "! ");
+//           setLoading(false);
+//         });
+//       return () => controller.abort();
+//     },
+//     deps ? [...deps] : []
+//   );
 
-      setLoading(true);
-      apiClient
-        .get<FetchResponse<T>>(endpoint, {
-          signal: controller.signal,
-          ...requestConfig,
-        })
-        .then((response) => {
-          setData(response.data.results);
-          setLoading(false);
-          console.log(response.data.results);
-        })
-        .catch((error) => {
-          if (error instanceof CanceledError) {
-            return;
-          }
-          setError(error.message + "! ");
-          setLoading(false);
-        });
-      return () => controller.abort();
-    },
-    deps ? [...deps] : []
-  );
+//   return { data, error, isLoading };
+// };
 
-  return { data, error, isLoading };
-};
-
-export default useData;
+// export default useData;
