+++ src/hooks/usePlatforms.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { FetchResponse } from "./useData";
+import { FetchResponse } from "../services/api-client";
 import apiClient from "../services/api-client";
 
 export interface Platform {
