+++ src/services/api-client.ts
@@ -1,8 +1,13 @@
 import axios from "axios";
 
+export interface FetchResponse<T> {
+  count: number;
+  results: T[];
+}
+
 export default axios.create({
-    baseURL: 'https://api.rawg.io/api',
-    params: {
-        key: '025c520084864e78b974363cfeac2127'
-    }
-})
+  baseURL: "https://api.rawg.io/api",
+  params: {
+    key: "025c520084864e78b974363cfeac2127",
+  },
+});
